+++ planck-cljs/src/planck/core.cljs
@@ -376,7 +376,7 @@
   "Returns the var to which a symbol will be resolved in the namespace, else
   nil."
   [ns sym]
-  (^:private-var-access-nowarn repl/ns-resolve ns sym))
+  (#'repl/ns-resolve ns sym))
 
 (s/fdef ns-resolve
   :args (s/cat :ns symbol? :sym symbol?)
@@ -386,7 +386,7 @@
   "Returns the var to which a symbol will be resolved in the current
   namespace, else nil."
   [sym]
-  (^:private-var-access-nowarn repl/resolve sym))
+  (#'repl/resolve sym))
 
 (s/fdef resolve
   :args (s/cat :sym symbol?)
@@ -398,9 +398,9 @@
   The namespace must exist. The var will adopt any metadata from the name
   symbol. Returns the var."
   ([ns name]
-   (^:private-var-access-nowarn repl/intern ns name))
+   (#'repl/intern ns name))
   ([ns name val]
-   (^:private-var-access-nowarn repl/intern ns name val)))
+   (#'repl/intern ns name val)))
 
 (s/fdef intern
   :args (s/cat :ns (s/or :sym symbol? :ns #(instance? Namespace %))
@@ -411,7 +411,7 @@
   [state ns]
   (-> state
     (assoc-in [:cljs.analyzer/namespaces ns]
-      (get-in (deref ^:private-var-access-nowarn repl/st) [:cljs.analyzer/namespaces ns]))))
+      (get-in @@#'repl/st [:cljs.analyzer/namespaces ns]))))
 
 (defn init-empty-state
   "An init function for use with cljs.js/empty-state which initializes the
@@ -443,5 +443,5 @@
 
 ;; Ensure planck.io and planck.http are loaded so that their
 ;; facilities are available
-(^:private-var-access-nowarn repl/side-load-ns 'planck.http)
-(^:private-var-access-nowarn repl/side-load-ns 'planck.io)
+(#'repl/side-load-ns 'planck.http)
+(#'repl/side-load-ns 'planck.io)